requirements.txt
tavily-python==0.3.3
ollama==0.2.1
python-dotenv==1.0.0